[#89] feat(frontend): implement premium Landing Page blocks and Hero overlay layout
